refactor: 모집 안내 페이지 컴포넌트 분리
